Add -kubermatic-delete-cluster flag, true by default (#1500)

+++ api/cmd/kubermatic-e2e/e2e_test_runner.go
@@ -77,7 +77,9 @@ func (ctl *e2eTestRunner) run(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
-	defer ctl.deleteCluster(cluster.Name)
+	if ctl.runOpts.DeleteCluster {
+		defer ctl.deleteCluster(cluster.Name)
+	}
 	log.Infof("created cluster named: %s", cluster.Name)
 
 	log.Info("waiting for cluster to become healthy")
